client_io_multiplexing

+++ common/tcp_client.h
@@ -9,5 +9,5 @@
 #ifndef _TCP_CLIENT_H
 #define _TCP_CLIENT_H
 int socket_connect(char *host, int port);
-int socket_connect_timeout(char *host, int port, int timeout);
+int socket_connect_timeout(char *host, int port, long timeout);
 #endif
